+++ app/keepers.go
@@ -50,12 +50,10 @@ import (
 	slashingtypes "github.com/cosmos/cosmos-sdk/x/slashing/types"
 	stakingkeeper "github.com/cosmos/cosmos-sdk/x/staking/keeper"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
-	packetforwardmiddleware "github.com/cosmos/ibc-apps/middleware/packet-forward-middleware/v8/packetforward"
 	packetforwardkeeper "github.com/cosmos/ibc-apps/middleware/packet-forward-middleware/v8/packetforward/keeper"
 	packetforwardtypes "github.com/cosmos/ibc-apps/middleware/packet-forward-middleware/v8/packetforward/types"
 	capabilitykeeper "github.com/cosmos/ibc-go/modules/capability/keeper"
 	capabilitytypes "github.com/cosmos/ibc-go/modules/capability/types"
-	ibctransfer "github.com/cosmos/ibc-go/v8/modules/apps/transfer"
 	ibctransferkeeper "github.com/cosmos/ibc-go/v8/modules/apps/transfer/keeper"
 	ibctransfertypes "github.com/cosmos/ibc-go/v8/modules/apps/transfer/types"
 	ibcporttypes "github.com/cosmos/ibc-go/v8/modules/core/05-port/types"
@@ -80,7 +78,6 @@ import (
 
 	appparams "github.com/dymensionxyz/dymension/v3/app/params"
 
-	"github.com/dymensionxyz/dymension/v3/x/bridgingfee"
 	delayedackmodule "github.com/dymensionxyz/dymension/v3/x/delayedack"
 	delayedackkeeper "github.com/dymensionxyz/dymension/v3/x/delayedack/keeper"
 	delayedacktypes "github.com/dymensionxyz/dymension/v3/x/delayedack/types"
@@ -507,38 +504,6 @@ func (a *AppKeepers) InitKeepers(
 		a.IBCKeeper.ChannelKeeper,
 		govModuleAddress,
 	)
-}
-
-func (a *AppKeepers) InitTransferStack() {
-	a.TransferStack = ibctransfer.NewIBCModule(a.TransferKeeper)
-	a.TransferStack = bridgingfee.NewIBCModule(
-		a.TransferStack.(ibctransfer.IBCModule),
-		*a.RollappKeeper,
-		a.DelayedAckKeeper,
-		a.TransferKeeper,
-		*a.TxFeesKeeper,
-	)
-	a.TransferStack = packetforwardmiddleware.NewIBCMiddleware(
-		a.TransferStack,
-		a.PacketForwardMiddlewareKeeper,
-		0,
-		packetforwardkeeper.DefaultForwardTransferPacketTimeoutTimestamp,
-	)
-
-	a.TransferStack = denommetadatamodule.NewIBCModule(a.TransferStack, a.DenomMetadataKeeper, a.RollappKeeper)
-	// already instantiated in SetupHooks()
-	a.DelayedAckMiddleware.Setup(
-		delayedackmodule.WithIBCModule(a.TransferStack),
-		delayedackmodule.WithKeeper(a.DelayedAckKeeper),
-		delayedackmodule.WithRollappKeeper(a.RollappKeeper),
-	)
-	a.TransferStack = a.DelayedAckMiddleware
-	a.TransferStack = genesisbridge.NewIBCModule(a.TransferStack, a.RollappKeeper, a.TransferKeeper, a.DenomMetadataKeeper)
-
-	// Create static IBC router, add transfer route, then set and seal it
-	ibcRouter := ibcporttypes.NewRouter()
-	ibcRouter.AddRoute(ibctransfertypes.ModuleName, a.TransferStack)
-	a.IBCKeeper.SetRouter(ibcRouter)
 }
 
 func (a *AppKeepers) SetupHooks() {
